Split page.tsx up into separate components and placed them into _components folder
