+++ app/_components/DeleteIcon.jsx
